chore: Modify test component script
